- fixed system.cpu.util[] (Eugene) - improved configuration script (Eugene) - fixed floats values (Eugene) - developed agent result convertion (Eugene) - more fixes
- TODO: WinXX agent \!\!\!

git-svn-id: svn://svn.zabbix.com/trunk@3886 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ src/zabbix_agent/interfaces.h
@@ -27,8 +27,8 @@ typedef struct s_single_interface_data
 {
 	char    *name;
 	int	clock[60*15];
-	float	sent[60*15];
-	float	received[60*15];
+	double	sent[60*15];
+	double	received[60*15];
 } ZBX_SINGLE_INTERFACE_DATA;
 
 typedef struct s_interfaces_data
@@ -46,8 +46,8 @@ INTERFACE
 {
 	char    *interface;
 	int	clock[60*15];
-	float	sent[60*15];
-	float	received[60*15];
+	double	sent[60*15];
+	double	received[60*15];
 };
 
 void	collect_stats_interfaces(FILE *outfile);
